requirements.txt
streamlit==1.52.1
openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.26.0
httpx>=0.27.0
